Save site option flags as boolean to prevent double sanitizing
There's a sanitize callback registered for flags with scope="site" which
converts boolean true/false values to "active" or "inactive". Because of
this, we need to make sure to update flags as a boolean, to avoid
"double-sanitizing" errors: trying to set a post to "inactive" would be
type-juggled to "true", and that would be converted to "active" by the
sanitize callback.